moved most work to separate base file, templatization
