+++ core/base/arrayPreconditioning/ArrayPreconditioning.cpp
@@ -1,0 +1,6 @@
+#include <ArrayPreconditioning.h>
+
+ttk::ArrayPreconditioning::ArrayPreconditioning() {
+  // inherited from Debug: prefix will be printed at the beginning of every msg
+  this->setDebugMsgPrefix("ArrayPreconditioning");
+}
