+++ core/base/arrayPreconditioning/CMakeLists.txt
@@ -1,0 +1,8 @@
+ttk_add_base_library(arrayPreconditioning
+  SOURCES
+    ArrayPreconditioning.cpp
+  HEADERS
+    ArrayPreconditioning.h
+  DEPENDS
+    common
+)
