+++ core/vtk/ttkArrayPreconditioning/ttk.module
@@ -5,4 +5,5 @@ SOURCES
 HEADERS
   ttkArrayPreconditioning.h
 DEPENDS
+  arrayPreconditioning
   ttkAlgorithm
