+++ core/vtk/ttkArrayPreconditioning/ttkArrayPreconditioning.h
@@ -25,14 +25,17 @@
 // VTK Includes
 #include <ttkAlgorithm.h>
 
-#ifdef TTK_ENABLE_MPI
-#include <mpi.h>
-#endif
 #include <vtkDataArraySelection.h>
 #include <vtkNew.h>
 
+// TTK Base Includes
+#include <ArrayPreconditioning.h>
+
 class TTKARRAYPRECONDITIONING_EXPORT ttkArrayPreconditioning
-  : public ttkAlgorithm {
+  : public ttkAlgorithm // we inherit from the generic ttkAlgorithm class
+  ,
+    protected ttk::ArrayPreconditioning // and we inherit from the base class
+{
 
 public:
   static ttkArrayPreconditioning *New();
@@ -59,15 +62,6 @@ protected:
   int RequestData(vtkInformation *request,
                   vtkInformationVector **inputVector,
                   vtkInformationVector *outputVector) override;
-
-#ifdef TTK_ENABLE_MPI
-  void ReceiveAndAddToVector(
-    MPI_Datatype mpi_values,
-    int rankFrom,
-    int structTag,
-    int intTag,
-    std::vector<std::vector<ttk::value>> &unsortedReceivedValues);
-#endif
 
 private:
   vtkNew<vtkDataArraySelection> ArraySelection{};
