+++ paraview/xmls/GhostCellPreprocessing.xml
@@ -35,7 +35,7 @@
 
 
       <Hints>
-        <ShowInMenu category="TTK - Scalar Data" />
+        <ShowInMenu category="TTK - Domain" />
       </Hints>
     </SourceProxy>
   </ProxyGroup>
